added usecase for repo operations and worker for data sync to backend

+++ app/src/main/java/com/bridge/androidtechnicaltest/pupil/data/datasources/local/PupilDao.kt
@@ -45,7 +45,7 @@ interface PupilDao {
     suspend fun deletePupilsWithSyncedStatus()
 
     @Query("DELETE FROM Pupils WHERE pupil_id = :pupilId")
-    suspend fun deleteByPupilId(pupilId: Long)
+    suspend fun deleteByPupilId(pupilId: Int)
 
     @Query("SELECT * FROM Pupils WHERE sync_status != 'SYNCED'")
     suspend fun getUnsyncedPupils(): List<LocalPupil>
